Add `Spree::EventBus`
We need an easier way to hook into events such as an order being
confirmed/cancelled, a carton getting shipped, et cetera.

Although "EventBus" may be a contentious name, the purpose of this class
is to "delegate" events that happen in the system to appropriate
processors. This is an in-process publish/subscribe model. To add
behaviour to run during during an event, a developer needs to create a
processor (see `Spree::Events::Processors::MailProcessor` as an
example) and subscribe it to the appropriate event class, such as
`Spree::Events::OrderConfirmedEvent`.