Add `Spree::EventBus`
We need an easier way to hook into events such as an order being
confirmed/cancelled, a carton getting shipped, et cetera.

The EventBus delegates events to appropriate processors (classes that
run procs, see the Spree::Events::Processors::MailProcessor as an
example). This is an in-process publish/subscribe model. To add
behaviour to run after an event, a developer should create a processor
and subscribe it to the appropriate event class, such as
`Spree::Events::OrderConfirmedEvent`.

See PR 2431 for an example of how this works in practice.