Use events in `Spree::Order`
Instead of calling ActionMailer methods directly from the Order code, we
can decouple it via publishing to the EventBus. This way stores can
subscribe to these events and create their own custom processors to
programmatically send/not send emails when different things happen in
Solidus.

See PR 2431 as a working example of using the EventBus.