Use events in `Spree::OrderCancellations`
Instead of calling ActionMailer methods directly from the cancellations
code, we can decouple it via publishing to the EventBus. This way stores
can subscribe to these events and create their own customer processors
to programmatically send/not send emails when different things happen in
Solidus.

See PR 2431 as a working example of using the EventBus.