Add MailProcessor
This class subscribes to events (via the Spree::EventBus) and runs its
defined methods when appropriate. Doing this decouples specific
ActionMailer calls from "events" (at the moment mostly Rails callbacks)
in Solidus. Developers can follow this class as an example on how to
create a processor for the Spree::EventBus.