Add suppression flag to `CartonShipppedEvent`
Since our events are not given live AR objects there is no way for us to
determine the value of the `suppress_mailer` option that was selected
when a carton was shipped.

We didn't want to persist this flag on a `Shipment`, and conditionally
using ActionMailer in the API controller is exactly what PR 2431 is
working to avoid.