Playing with Activesupport Notifications
This is a very basic example of how Activesupport notifications can be used to
generate events in Solidus.

The example is quite rough and is meant to show how simple a basic integration
actually is.

I think that we may encapsulate `ActiveSupport::Notifications.subscribe` and
`ActiveSupport::Notifications.instrument` into a custom class so that, if in
the future we want to switch the notification backend system, this can be done
easily without requiring to replace in lots of files all the occurrences of
`ActiveSupport::Notifications.subscribe` and similar.